Make the plugin enabled by default after install
I wondered why it didn't work after installation for a while because I know plugins do not need to be enabled after installation at deadbeef level. Hopefully this will make it more friendly to newcomers.

Also added an include on `<string.h>` so that compiler won't complain about the implicit definition of `strcmp`.